requirements.txt
django>=4.2.0
tensorflow>=2.12.0
pillow>=9.5.0
numpy>=1.24.0 